Split extension task in to V1 and V2
