+++ .github/workflows/extension.yml
@@ -60,7 +60,7 @@ jobs:
         working-directory: '${{ github.workspace }}/extension'
 
       - name: Build
-        run: npm run build:prod
+        run: npm run build
         working-directory: '${{ github.workspace }}/extension'
 
       - name: Install tfx-cli
@@ -74,11 +74,11 @@ jobs:
           MAJOR_MINOR_PATCH: ${{ steps.gitversion.outputs.majorMinorPatch }}
           BUILD_NUMBER: ${{ github.run_number }}
 
-      - name: Update values in extension/task/task.json
+      - name: Update values in extension/tasks/dependabot/dependabotV2/task.json
         run: |
-          echo "`jq '.version.Major=${{ steps.gitversion.outputs.major }}' extension/task/task.json`" > extension/task/task.json
-          echo "`jq '.version.Minor=${{ steps.gitversion.outputs.minor }}' extension/task/task.json`" > extension/task/task.json
-          echo "`jq '.version.Patch=${{ github.run_number }}' extension/task/task.json`" > extension/task/task.json
+          echo "`jq '.version.Major=${{ steps.gitversion.outputs.major }}' extension/tasks/dependabot/dependabotV2/task.json`" > extension/tasks/dependabot/dependabotV2/task.json
+          echo "`jq '.version.Minor=${{ steps.gitversion.outputs.minor }}' extension/tasks/dependabot/dependabotV2/task.json`" > extension/tasks/dependabot/dependabotV2/task.json
+          echo "`jq '.version.Patch=${{ github.run_number }}' extension/tasks/dependabot/dependabotV2/task.json`" > extension/tasks/dependabot/dependabotV2/task.json
 
       - name: Create Extension (dev)
         run: >
