+++ docs/extension.md
@@ -34,7 +34,7 @@ npm run build:prod
 To generate the  Azure DevOps `.vsix` extension package for testing, you'll first need to [create a publisher account](https://learn.microsoft.com/en-us/azure/devops/extend/publish/overview?view=azure-devops#create-a-publisher) on the [Visual Studio Marketplace Publishing Portal](https://marketplace.visualstudio.com/manage/createpublisher?managePageRedirect=true). After this, override your publisher ID below and generate the extension with:
 
 ```bash
-npx tfx-cli extension create --overrides-file overrides.local.json --override "{\"publisher\": \"your-publisher-id-here\"}" --json5
+npm run package -- --overrides-file overrides.local.json --override "{\"publisher\": \"your-publisher-id-here\"}"
 ```
 
 ## Installing the extension
