+++ extension/overrides.local.json
@@ -1,5 +1,5 @@
 {
   "id": "dependabot-local",
-  "version": "0.1.0.6",
+  "version": "0.2.0.0",
   "name": "Dependabot (Local)"
 }
