+++ extension/tasks/dependabot/dependabotV1/icon.png
