+++ extension/tasks/dependabot/dependabotV1/index.ts
@@ -1,9 +1,9 @@
 import * as tl from 'azure-pipelines-task-lib/task';
 import { ToolRunner } from 'azure-pipelines-task-lib/toolrunner';
-import { IDependabotRegistry, IDependabotUpdate } from './IDependabotConfig';
-import getSharedVariables from './utils/getSharedVariables';
-import { parseConfigFile } from './utils/parseConfigFile';
-import { resolveAzureDevOpsIdentities } from './utils/resolveAzureDevOpsIdentities';
+import { IDependabotRegistry, IDependabotUpdate } from '../../utils/IDependabotConfig';
+import getSharedVariables from '../../utils/getSharedVariables';
+import { parseConfigFile } from '../../utils/parseConfigFile';
+import { resolveAzureDevOpsIdentities } from '../../utils/resolveAzureDevOpsIdentities';
 
 async function run() {
   try {
