+++ extension/tasks/dependabot/dependabotV2/index.ts
@@ -1,0 +1,17 @@
+import { setResult, TaskResult } from "azure-pipelines-task-lib/task"
+import { debug, warning, error } from "azure-pipelines-task-lib/task"
+
+async function run() {
+  try {
+    
+    // TODO: This...
+    setResult(TaskResult.Succeeded);
+
+  }
+  catch (e: any) {
+      error(`Unhandled exception: ${e}`);
+      setResult(TaskResult.Failed, e?.message);
+  }
+}
+
+run();
