+++ extension/tasks/utils/IDependabotConfig.ts
