+++ extension/tasks/utils/getDockerImageTag.ts
