+++ extension/tasks/utils/getSharedVariables.ts
