+++ extension/vss-extension.json
@@ -31,7 +31,7 @@
   },
   "files": [
     {
-      "path": "task"
+      "path": "tasks"
     },
     {
       "path": "images",
@@ -44,7 +44,7 @@
       "type": "ms.vss-distributed-task.task",
       "targets": ["ms.vss-distributed-task.tasks"],
       "properties": {
-        "name": "task"
+        "name": "tasks/dependabot"
       }
     }
   ]
